+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ----------------------
 ##### Fixed
 - Download links in file input and draw widgets not working in IE11.
+- The required asterisks (*) do not update in real-time.
 
 [4.55.1] - 2018-07-25
 ----------------------